+++ DFSA/src/estimadores/Estimador.java
@@ -4,7 +4,7 @@ public abstract class Estimador {
 	
 	protected int n_til;
 	
-	public abstract int backlog( int e, int s, int c);
+	public abstract int backlog( double e, double s, double c);
 	
 	public int frameLenght(int s) {
 		return this.n_til - s;
